Move FlipNegatedTernaryInstanceofRector to instanceof rules set
