+++ rules-tests/Instanceof_/Rector/Ternary/FlipNegatedTernaryInstanceofRector/Fixture/skip_no_cond.php.inc
