+++ rules-tests/Instanceof_/Rector/Ternary/FlipNegatedTernaryInstanceofRector/Fixture/some_class.php.inc
