+++ rules-tests/Instanceof_/Rector/Ternary/FlipNegatedTernaryInstanceofRector/config/configured_rule.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
 
-use Rector\TypeDeclaration\Rector\Ternary\FlipNegatedTernaryInstanceofRector;
+use Rector\Instanceof_\Rector\Ternary\FlipNegatedTernaryInstanceofRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(FlipNegatedTernaryInstanceofRector::class);
